+++ src/app/dataset-block/metadata-block/block.service.ts
@@ -10,21 +10,25 @@ import { Observable } from "rxjs";
 import { map } from "rxjs/operators";
 import { DatasetApi } from "src/app/api/dataset.api";
 import {
+    AddPushSource,
     GetMetadataBlockQuery,
     MetadataBlockFragment,
     MetadataEventType,
     MetadataManifestFormat,
 } from "src/app/api/kamu.graphql.interface";
-import { MaybeUndefined } from "src/app/interface/app.types";
+import { MaybeNull, MaybeUndefined } from "src/app/interface/app.types";
 import { DatasetInfo } from "src/app/interface/navigation.interface";
 import { MetadataBlockInfo } from "./metadata-block.types";
 import { MetadataBlockExtended } from "./../../api/kamu.graphql.interface";
+import { NavigationService } from "src/app/services/navigation.service";
 
 @Injectable({
     providedIn: "root",
 })
 export class BlockService {
+    public sourceNames: string[] = [];
     private datasetApi = inject(DatasetApi);
+    private navigationService = inject(NavigationService);
 
     public requestMetadataBlock(info: DatasetInfo, blockHash: string): Observable<MaybeUndefined<MetadataBlockInfo>> {
         return this.datasetApi.getBlockByHash({ ...info, blockHash }).pipe(
@@ -51,29 +55,53 @@ export class BlockService {
         accountName: string;
         datasetName: string;
         encoding: MetadataManifestFormat;
-    }): Observable<MetadataBlockExtended[]> {
+    }): Observable<MaybeNull<string>> {
         return this.datasetApi
             .getBlocksByEventType({ ...params, eventTypes: [MetadataEventType.SetPollingSource] })
-            .pipe(map((data) => data.datasets.byOwnerAndName?.metadata.metadataProjection as MetadataBlockExtended[]));
+            .pipe(
+                map((data) => {
+                    const blocks = data.datasets.byOwnerAndName?.metadata.metadataProjection as MetadataBlockExtended[];
+                    return blocks.length ? (blocks[0].encoded?.content as string) : null;
+                }),
+            );
     }
 
     public requestBlocksBySetTransformEvent(params: {
         accountName: string;
         datasetName: string;
         encoding: MetadataManifestFormat;
-    }): Observable<MetadataBlockExtended[]> {
-        return this.datasetApi
-            .getBlocksByEventType({ ...params, eventTypes: [MetadataEventType.SetTransform] })
-            .pipe(map((data) => data.datasets.byOwnerAndName?.metadata.metadataProjection as MetadataBlockExtended[]));
+    }): Observable<MaybeNull<string>> {
+        return this.datasetApi.getBlocksByEventType({ ...params, eventTypes: [MetadataEventType.SetTransform] }).pipe(
+            map((data) => {
+                const blocks = data.datasets.byOwnerAndName?.metadata.metadataProjection as MetadataBlockExtended[];
+                return blocks.length ? (blocks[0].encoded?.content as string) : null;
+            }),
+        );
     }
 
     public requestBlocksByAddPushSourceEvent(params: {
         accountName: string;
         datasetName: string;
+        sourceName: string;
         encoding: MetadataManifestFormat;
-    }): Observable<MetadataBlockExtended[]> {
-        return this.datasetApi
-            .getBlocksByEventType({ ...params, eventTypes: [MetadataEventType.AddPushSource] })
-            .pipe(map((data) => data.datasets.byOwnerAndName?.metadata.metadataProjection as MetadataBlockExtended[]));
+    }): Observable<MaybeNull<string>> {
+        return this.datasetApi.getBlocksByEventType({ ...params, eventTypes: [MetadataEventType.AddPushSource] }).pipe(
+            map((data) => {
+                const blocks = data.datasets.byOwnerAndName?.metadata.metadataProjection as MetadataBlockExtended[];
+                this.sourceNames = blocks.map((item) => (item.event as AddPushSource).sourceName);
+                if (params.sourceName) {
+                    if (!this.sourceNames.includes(params.sourceName)) {
+                        this.navigationService.navigateToPageNotFound();
+                    }
+                    const block = blocks.filter(
+                        (item) =>
+                            item.event.__typename === "AddPushSource" && item.event.sourceName === params.sourceName,
+                    )[0];
+                    return block.encoded?.content as string;
+                } else {
+                    return null;
+                }
+            }),
+        );
     }
 }
